+++ src/event.rs
@@ -1,5 +1,4 @@
 use crate::{
-    input::{Key, MouseButton},
     spatial::{Position, Size},
     timeout::{TimeoutId, TimeoutTag},
 };
@@ -102,4 +101,42 @@ impl MouseEvent {
             }
         }
     }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+#[non_exhaustive]
+#[cfg_attr(
+    feature = "serde",
+    derive(serde::Serialize, serde::Deserialize),
+    serde(rename_all = "camelCase")
+)]
+pub enum MouseButton {
+    Right,
+    Middle,
+    Left,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+#[non_exhaustive]
+#[cfg_attr(
+    feature = "serde",
+    derive(serde::Serialize, serde::Deserialize),
+    serde(rename_all = "camelCase")
+)]
+pub enum Key {
+    Left,
+    Right,
+    Down,
+    Up,
+    Return,
+    Backspace,
+    Delete,
+    Shift,
+    Ctrl,
+    Alt,
+    Tab,
+    Esc,
+    Char(char),
+    #[cfg_attr(feature = "serde", serde(other))]
+    Other,
 }