+++ CHANGES.md
@@ -11,6 +11,7 @@ drop
 core
 ----
 
+- [CHANGE] Merge `input` module into `event` module
 - [ADD] Add `println!()`, `eprintln!()` and `dbg!()` for WebAssembly build
 - [CHANGE] Remove `log` crate from dependencies
 - [CHANGE] Remove `System::console_log`